Seprate test install and no install.
