+++ tests/phpunit/Fixtures/command/install_no_config_file/base/composer.json
@@ -1,0 +1,19 @@
+{
+    "name": "yourorg/yourtempaltepackage",
+    "description": "Your template package description",
+    "type": "project",
+    "license": "proprietary",
+    "require": {
+        "php": ">=8.2",
+        "monolog/monolog": "^2.0"
+    },
+    "require-dev": {
+        "alexskrypnyk/customizer": "^0.2"
+    },
+    "minimum-stability": "dev",
+    "config": {
+        "allow-plugins": {
+            "alexskrypnyk/customizer": true
+        }
+    }
+}
